refactor: tdd mem complete
