+++ assignments/api-express/mlopsapi/mem-db/user.json
@@ -1,16 +1,16 @@
 [
-	{
-		"id": 2,
-		"name": "sm2",
-		"age": 67
-	},
-	{
-		"id": 3,
-		"name": "seokmin"
-	},
-	{
-		"id": 4,
-		"name": "sm",
-		"age": 26
-	}
+  {
+    "id": 2,
+    "name": "sm2",
+    "age": 67
+  },
+  {
+    "id": 3,
+    "name": "seokmin"
+  },
+  {
+    "id": 4,
+    "name": "sm",
+    "age": 26
+  }
 ]
